Fixing types.

+++ integrations/Unity/simenv-unity/Assets/RLAgents/AgentManager.cs
@@ -18,7 +18,7 @@ namespace SimEnv.Agents {
         }
 
         float physicsUpdateRate;
-        float frameSkip;
+        int frameSkip;
 
         public AgentManager() {
